Complete end to end create flow

+++ service/src/main/java/io/github/aj8gh/skeleton/service/model/Skeleton.java
@@ -11,7 +11,7 @@ public class Skeleton {
 
   UUID id;
   String name;
-  long bones;
+  int bones;
   Instant createdAt;
   Instant updatedAt;
 }
